pages,modules shared static

+++ Logging/Models/FileDataProvider.cs
@@ -112,6 +112,7 @@ namespace YetaWF.Modules.Logging.DataProvider.File {
         }
         public async Task<bool> InstallModelAsync(List<string> errorList) {
             if (YetaWF.Core.Log.Logging.DefinedLoggerType != typeof(LogRecordDataProvider)) return true;
+            if (YetaWF.Core.IO.Caching.MultiInstance) throw new InternalError("Installing new models is not possible when distributed caching is enabled");
             if (_isInstalled == null)
                 _isInstalled = Directory.Exists(Path.GetDirectoryName(LogFile));
             await YetaWF.Core.Log.Logging.SetupLoggingAsync();
@@ -122,6 +123,7 @@ namespace YetaWF.Modules.Logging.DataProvider.File {
 
         public Task<bool> UninstallModelAsync(List<string> errorList) {
             if (YetaWF.Core.Log.Logging.DefinedLoggerType != typeof(LogRecordDataProvider)) return Task.FromResult(true);
+            if (YetaWF.Core.IO.Caching.MultiInstance) throw new InternalError("Adding site data is not possible when distributed caching is enabled");
             YetaWF.Core.Log.Logging.TerminateLogging();
             try {
                 System.IO.File.Delete(LogFile);
